src/tbcxr_model01/function/requirements.txt
azure-functions
pillow
torch
torchvision